+++ src/components/slotmachine/components/paytable.tsx
