+++ src/components/slotmachine/data.tsx
